Put Aphrodite Example in line with other examples
